pagina principal de pagos y navegacion base
Base para la pagina de pagos y su navegacion entre los diferentes pasos 
sin css
